Moved the default css generation function into `html_generator.rs`
