+++ src/html_generator.rs
@@ -144,3 +144,208 @@ fn format_title(file_name: &str) -> String {
         .collect::<Vec<String>>()
         .join(" ")
 }
+
+/// Generates a default CSS stylesheet as a string.
+pub fn generate_default_css() -> String {
+    r#"
+    body {
+    background-color: #121212;
+    color: #e0e0e0;
+    font-family:
+        -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu,
+        Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
+    line-height: 1.75;
+    margin: 0;
+    padding: 0;
+    }
+
+    /* Card-like container for the page content */
+    #content {
+    background-color: #1e1e1e;
+    max-width: 780px;
+    margin: 1.5rem;
+    padding: 2rem;
+    border-radius: 12px;
+    box-shadow: 0 0 0 1px #2c2c2c;
+    }
+
+    header {
+    background-color: #1a1a1a;
+    border-bottom: 1px solid #333;
+    position: sticky;
+    top: 0;
+    z-index: 1000;
+    }
+
+    nav {
+    padding: 1rem 2rem;
+    display: flex;
+    justify-content: center;
+    }
+
+    nav ul {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    gap: 1rem;
+    }
+
+    nav ul li {
+    margin: 0;
+    }
+
+    nav ul li a {
+    color: #ddd;
+    text-decoration: none;
+    padding: 0.5rem 1rem;
+    border-radius: 6px;
+    transition: background-color 0.2s ease, color 0.2s ease;
+    }
+
+    nav ul li a:hover {
+    background-color: #2f2f2f;
+    color: #fff;
+    }
+
+    nav ul li a.active {
+    background-color: #4ea1f3;
+    color: #121212;
+    }
+    h1,
+    h2,
+    h3,
+    h4,
+    h5,
+    h6 {
+    color: #ffffff;
+    line-height: 1.3;
+    margin-top: 2rem;
+    margin-bottom: 1rem;
+    }
+
+    h1 {
+    font-size: 2.25rem;
+    border-bottom: 2px solid #2c2c2c;
+    padding-bottom: 0.3rem;
+    }
+    h2 {
+    font-size: 1.75rem;
+    border-bottom: 1px solid #2c2c2c;
+    padding-bottom: 0.2rem;
+    }
+    h3 {
+    font-size: 1.5rem;
+    }
+    h4 {
+    font-size: 1.25rem;
+    }
+    h5,
+    h6 {
+    font-size: 1rem;
+    font-weight: normal;
+    }
+
+    p {
+    margin-bottom: 1.2rem;
+    }
+
+    a {
+    color: #4ea1f3;
+    text-decoration: none;
+    transition: color 0.2s ease-in-out;
+    }
+    a:hover {
+    color: #82cfff;
+    text-decoration: underline;
+    }
+
+    img {
+    max-width: 100%;
+    height: auto;
+    display: block;
+    margin: 1.5rem auto;
+    border-radius: 8px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.3);
+    }
+
+    pre {
+    background-color: #2a2a2a;
+    padding: 1rem;
+    border-radius: 8px;
+    overflow-x: auto;
+    font-size: 0.9rem;
+    box-shadow: inset 0 0 0 1px #333;
+    }
+    code {
+    font-family: SFMono-Regular, Consolas, "Liberation Mono", Menlo, monospace;
+    background-color: #2a2a2a;
+    padding: 0.2em 0.4em;
+    border-radius: 4px;
+    font-size: 0.95em;
+    color: #dcdcdc;
+    }
+
+    blockquote {
+    border-left: 4px solid #555;
+    padding-left: 1rem;
+    color: #aaa;
+    font-style: italic;
+    margin: 1.5rem 0;
+    background-color: #1a1a1a;
+    border-radius: 4px;
+    }
+
+    ul,
+    ol {
+    padding-left: 1.5rem;
+    margin-bottom: 1.2rem;
+    }
+    li {
+    margin-bottom: 0.5rem;
+    }
+
+    table {
+    width: 100%;
+    border-collapse: collapse;
+    margin: 2rem 0;
+    background-color: #1e1e1e;
+    border: 1px solid #333;
+    border-radius: 8px;
+    overflow: hidden;
+    font-size: 0.95rem;
+    }
+
+    th,
+    td {
+    padding: 0.75rem 1rem;
+    text-align: left;
+    border-bottom: 1px solid #333;
+    }
+
+    th {
+    background-color: #2a2a2a;
+    color: #ffffff;
+    font-weight: 600;
+    }
+
+    tr:nth-child(even) td {
+    background-color: #222;
+    }
+
+    tr:hover td {
+    background-color: #2f2f2f;
+    }
+
+    td {
+    color: #ddd;
+    }
+
+    hr {
+    border: none;
+    border-top: 1px solid #333;
+    margin: 2rem 0;
+    }
+    "#
+    .to_string()
+}
